move AbstractConfigFilter
